+++ rover_auxiliary/src/BMS/bms_node.hpp
@@ -44,7 +44,7 @@ class BMSDataNode : public rclcpp::Node
     void callbackBMSData(void);
     bool getData(void);
     template<size_t N>
-    bool parse(std::string_view view_, std::string_view expectedPrefix_, std::array<uint16_t, N>& dataArray_);
+    bool parse(std::string_view view_, std::string_view expectedPrefix_, std::array<int32_t, N>& dataArray_);
     std::optional<std::string> readDataFrame();
 
     std::array<int32_t, AMP_DATA_TYPES> _ampArray{};
